+++ bitorch/layers/__init__.py
@@ -16,9 +16,11 @@ from .qconv1d import QConv1d, QConv1d_NoAct
 from .qconv2d import QConv2d, QConv2d_NoAct
 from .qconv3d import QConv3d, QConv3d_NoAct
 from .qlinear import QLinear
+from .pact import Pact
+from .qembedding import QEmbedding, QEmbeddingBag
 
 __all__ = [
     "InputGraphicalDebug", "InputPrintDebug", "WeightGraphicalDebug", "WeightPrintDebug",
     "ShapePrintDebug", "QActivation", "QConv1d", "QConv2d", "QConv3d", "QConv1d_NoAct",
-    "QConv2d_NoAct", "QConv3d_NoAct", "QLinear"
+    "QConv2d_NoAct", "QConv3d_NoAct", "QLinear", "QEmbedding", "QEmbeddingBag", "Pact",
 ]